+++ changelog.md
@@ -4,6 +4,8 @@
 
 - **Fixed:**
   - Command history: Add text ... overflow systems for long commands to prevent drop up display from being hidden off screen
+  - Code editor:
+    - Virtual/area editor: Fix issue with room preview and item formatting when item text is empty
 - **Changed:**
   - Update drivelist 7.0.0 to 7.0.1
   - Update electron 4.0.2 to 4.0.5